shuffled test stuff around, collected common stuff in __init__
