+++ tests/compiler.py
@@ -14,18 +14,14 @@
 
 
 """
-unittest for sibilant.compile
+Some common unit test utility funtions
 
 author: Christopher O'Brien  <obriencj@gmail.com>
 license: LGPL v.3
 """
 
 
-import dis
-
-from contextlib import contextmanager
 from fractions import Fraction as fraction
-from functools import partial
 from unittest import TestCase
 
 from sibilant import (
@@ -40,105 +36,9 @@ from sibilant.compiler import (
     CodeFlag,
 )
 
-from sibilant.module import (
-    fake_module_from_env, init_module, load_module_1,
-    run_time,
+from . import (
+    Object, compile_expr, compile_expr_no_tco,
 )
-
-from sibilant.parse import source_str
-
-
-class Object(object):
-    pass
-
-
-def compile_expr(src_str, **base):
-    mod = fake_module_from_env(base)
-    init_module(mod, source_str(src_str, "<unittest>"), None)
-
-    partial_run_time = partial(partial, run_time)
-
-    result = load_module_1(mod, run_time=partial_run_time)
-
-    return result, mod.__dict__
-
-
-def compile_expr_no_tco(src_str, **base):
-    mod = fake_module_from_env(base)
-
-    params = {"tco_enabled": False}
-
-    init_module(mod, source_str(src_str, "<unittest>"), None,
-                compiler_factory_params=params)
-
-    partial_run_time = partial(partial, run_time)
-
-    result = load_module_1(mod, run_time=partial_run_time)
-
-    return result, mod.__dict__
-
-
-def compile_dis_expr(src_str, **base):
-    mod = fake_module_from_env(base)
-    init_module(mod, source_str(src_str, "<unittest>"), None)
-
-    code_objs = []
-    def partial_run_time(module, code_obj):
-
-        dis.show_code(code_obj)
-        print("Disassembly:")
-        dis.dis(code_obj)
-
-        code_objs.append(code_obj)
-        return partial(run_time, module, code_obj)
-
-    result = load_module_1(mod, run_time=partial_run_time)
-
-    return result, mod.__dict__
-
-
-def make_accumulator():
-    accu = list()
-
-    def accumulate(x):
-        accu.append(x)
-        return x
-
-    return accu, accumulate
-
-
-def make_raise_accumulator(excclass=Exception):
-    accu = list()
-
-    def accumulate(x):
-        accu.append(x)
-        raise excclass(x)
-
-    return accu, accumulate
-
-
-def make_manager():
-    accumulator = list()
-
-    def accu(val):
-        accumulator.append(val)
-        return val
-
-    class Manager():
-        def __init__(self, initial, enter, leave):
-            self.enter = enter
-            self.leave = leave
-            accumulator.append(initial)
-
-        def __enter__(self):
-            accumulator.append(self.enter)
-            return accu
-
-        def __exit__(self, _a, _b, _c):
-            accumulator.append(self.leave)
-            return True
-
-    return accumulator, Manager
 
 
 class TestCompiler(TestCase):
