+++ tests/specials.py
@@ -27,16 +27,16 @@ from functools import partial
 from types import GeneratorType as generator
 from unittest import TestCase
 
-from .compiler import (
-    compile_expr, compile_dis_expr,
-    make_accumulator, make_raise_accumulator,
-    make_manager,
-)
-
 from sibilant import symbol, cons, nil, is_nil
 from sibilant.compiler import (
     Special, is_special,
     Macro, is_macro,
+)
+
+from . import (
+    compile_expr, compile_dis_expr,
+    make_accumulator, make_raise_accumulator,
+    make_manager,
 )
